requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
SQLAlchemy==2.0.29
email_validator==2.1.1
python-jose==3.3.0
numpy==1.26.4
pandas==2.2.1
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2
requests==2.31.0